+++ src/components/resultsComponent/AllResults.js
@@ -1,4 +1,8 @@
 import React from 'react'
+import {Col, Row, Button, Container, Navbar} from 'react-bootstrap'
+import aos from 'aos'
+import 'aos/dist/aos.css'
+
 import ShowsCount from './ShowsCount'
 import ViewCount from './ViewCount'
 import GenresCount from './GenresCount'
@@ -6,10 +10,9 @@ import PopularityCount from './PopularityCount'
 import RegionsCount from './RegionsCount'
 import RunTime from './RunTime'
 import Typewriter from 'typewriter-effect'
-import {Col, Row, Button, Container, Navbar} from 'react-bootstrap'
 import Summary from './Summary'
-import aos from 'aos'
-import 'aos/dist/aos.css'
+import RunTimeFunFact from "./FunFacts/RunTimeFunFact"
+import GenreFunFacts from "./FunFacts/GenreFunFacts"
 
 export class AllResults extends React.Component {
   constructor() {
@@ -25,8 +28,6 @@ export class AllResults extends React.Component {
     this.handleDoneTyping = this.handleDoneTyping.bind(this)
     this.handleShowResult = this.handleShowResult.bind(this)
     this.copyToClipboard = this.copyToClipboard.bind(this)
-    this.getRuntimeFunFact = this.getRuntimeFunFact.bind(this)
-    this.getGenreFunFact = this.getGenreFunFact.bind(this)
   }
 
   componentDidMount() {
@@ -46,7 +47,6 @@ export class AllResults extends React.Component {
   }
   handleShowResult() {
     this.setState({showResult: true})
-    window.scrollTo(0, 500)
   }
   copyToClipboard() {
     let dummy = document.createElement('input')
@@ -59,61 +59,6 @@ export class AllResults extends React.Component {
     document.execCommand('copy')
     document.body.removeChild(dummy)
     this.setState({copied: true})
-  }
-  getRuntimeFunFact() {
-    const sumMin = Object.keys(this.state.result.runtime.data).reduce(
-      (acc, key) => acc + this.state.result.runtime.data[key],
-      0
-    )
-    const sumHrs = Math.floor(sumMin / 60)
-    if (sumHrs < 24) return `called your grandma and made her day.`
-    if (sumHrs < 24 * 3)
-      return `fermented kimchi so you could eat something besides takeout and TV dinners for once.`
-    if (sumHrs < 24 * 7)
-      return `driven from New York City to Los Angeles. Or just, you know, go outside.`
-    if (sumHrs < 24 * 20)
-      return `grown some bok choy. Oh, you don't know what that is? Not surprising.`
-    if (sumHrs < 24 * 45)
-      return `lived an entire life… as a fruit fly. They come out of an egg, grow up, find a mate, lay eggs, lay more eggs, and die fulfilled in one month. Can you really say you’ve done better?`
-    if (sumHrs < 24 * 30 * 6) return `taken an entire semester of classes. Have you considered that? Improving yourself and learning new skills? Or are you content with your miserable existence?`
-    if (sumHrs < 24 * 30 * 3)
-      return `become a licensed real estate broker. Maybe you could have bought some property yourself and finally move out of your parent’s basement.`
-    return `had an entire pregnancy but I guess it’s understandable no one would trust you with a child much less yourself.`
-  }
-  getGenreFunFact() {
-    const resultsArr = Object.entries(this.state.result.genres)
-    //sort the data to large to small
-    const allSorted = resultsArr.sort((a, b) => b[1] - a[1])
-    const genresArr = allSorted.map(data => data[0])
-    if (genresArr[0] === 'Crime')
-      return (
-        <p>
-          SNL wrote a{' '}
-          <a href="https://www.youtube.com/watch?v=J4RdcE6H4Gs">sketch</a> about
-          people like you.
-        </p>
-      )
-    if (genresArr[0] === 'Adventure')
-      return `As of September 2020, adventure movies were the most popular movie genre in North America, with a total box office revenue of 63.57 billion U.S. dollars. How much of that amount was you?`
-    if (genresArr[0] === 'Animation')
-      return `The highest grossing animated film to date is The Lion King (2019).`
-    if (genresArr[0] === 'Comedy')
-      return (
-        <p>
-          There is something to be said that your favorite genre is comedy. Have
-          you heard of the{' '}
-          <a href="https://en.wikipedia.org/wiki/Sad_clown_paradox">
-            Sad Clown Paradox
-          </a>? 🤡
-        </p>
-      )
-    if (genresArr[0] === 'Sci-Fi')
-      return `The highest grossing science-fiction film to date is Avatar.`
-    if (genresArr[0] === 'Action')
-      return `You probably enjoy action because you're getting none.`
-    if (genresArr[0] === 'Reality-TV')
-      return `Reality-TV's title of "reality" is often criticized as being inaccurate because of claims that the genre frequently includes elements such as premeditated scripting (including a practice called "soft-scripting"), acting, urgings from behind-the-scenes crew to create specified situations of adversity and drama, and misleading editing. But you would know a thing or two about being fake.`
-    return `As of September 2020, adventure movies were the most popular movie genre in North America, with a total box office revenue of 63.57 billion U.S. dollars.`
   }
 
   render() {
@@ -177,7 +122,9 @@ export class AllResults extends React.Component {
                 </Col>
               </Row>
               <div>
-                  <p><strong style={{color: 'red'}}> Fun Fact: </strong> With the time you spent on Netflix, you could have {this.getRuntimeFunFact()}</p>
+                  <p>
+                    <strong style={{color: 'red'}}> Fun Fact: </strong> With the time you spent on Netflix, you could have <RunTimeFunFact result={this.state.result} />
+                    </p>
               </div>
               <Row className="justify-content-center">
                 <Col data-aos="zoom-out">
@@ -185,7 +132,10 @@ export class AllResults extends React.Component {
                 </Col>
               </Row>
               <div>
-                  <p><strong style={{color: 'red'}}> Fun Fact: </strong> {this.getGenreFunFact()}</p>
+                  <p>
+                    <strong style={{color: 'red'}}> Fun Fact: </strong>
+                    <GenreFunFacts result={this.state.result}/>
+                    </p>
               </div>
               <Row className="justify-content-center">
                 <Col data-aos="zoom-out">
